+++ challenges/BIRC/cb_2/src/client_backend.cc
@@ -53,7 +53,7 @@ User *FindUser(const char *name, unsigned short length, int *idx) {
 
 namespace Backend {
 
-void InitDataStores(int secret_page_i) {
+void InitDataStores(size_t secret_page_i) {
     g_secret = (const int *)secret_page_i;
 }
